Moving jmx itests package under as component - fixing PR-598
