+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/META-INF/MANIFEST.MF
@@ -1,8 +1,8 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: %BundleName
-Bundle-SymbolicName: org.jboss.tools.jmx.ui.bot.test
-Bundle-Version: 1.10.100.qualifier
+Bundle-SymbolicName: org.jboss.tools.as.jmx.ui.bot.itests
+Bundle-Version: 3.6.2.qualifier
 Bundle-Vendor: Red Hat
 Require-Bundle: org.junit,
  org.eclipse.reddeer.go;bundle-version="[2.1.0,3.0.0)",
