+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/build.properties
