+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/plugin.properties
