+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/pom.xml
@@ -1,27 +1,26 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" 
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.jboss.tools.jmx</groupId>
-    <artifactId>itests</artifactId>
-    <version>1.10.6-SNAPSHOT</version>
-  </parent>
-  <groupId>org.jboss.tools.jmx.tests</groupId>
-  <artifactId>org.jboss.tools.jmx.ui.bot.test</artifactId>
-  <version>1.10.100-SNAPSHOT</version>
-  <packaging>eclipse-test-plugin</packaging>
-  
-	 <properties>
-	 	<maven.test.failure.ignore>true</maven.test.failure.ignore>
-	 	<rd.config>${project.build.directory}/classes/requirements.json</rd.config>
+
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>org.jboss.tools.as</groupId>
+		<artifactId>itests</artifactId>
+		<version>3.6.2-SNAPSHOT</version>
+	</parent>
+	<groupId>org.jboss.tools.as.itests</groupId>
+	<artifactId>org.jboss.tools.as.jmx.ui.bot.itests</artifactId>
+	<packaging>eclipse-test-plugin</packaging>
+
+	<properties>
+		<maven.test.failure.ignore>true</maven.test.failure.ignore>
+		<rd.config>${project.build.directory}/classes/requirements.json</rd.config>
 		<systemProperties>-Drd.config=${rd.config} -Dusage_reporting_enabled=false -Djbosstools.test.wildfly.14.home=${jbosstools.test.wildfly.14.home}</systemProperties>
 		<surefire.timeout>14400</surefire.timeout>
 		<jbosstools.test.wildfly.14.home>${requirementsDirectory}/wildfly-14.0.0.Final</jbosstools.test.wildfly.14.home>
-		<suiteClass>org.jboss.tools.jmx.ui.bot.test.JMXAllTests</suiteClass>
+		<suiteClass>org.jboss.tools.as.jmx.ui.bot.itests.JMXAllTests</suiteClass>
 	</properties>
-	
+
 	<build>
 		<plugins>
 			<plugin>
@@ -49,21 +48,22 @@
 				</executions>
 			</plugin>
 			<plugin>
-	            <groupId>org.apache.maven.plugins</groupId>
-	            <artifactId>maven-compiler-plugin</artifactId>
-	            <version>3.5.1</version>
-	            <configuration>
-	                <source>1.8</source>
-	                <target>1.8</target>
-	            </configuration>
-        	</plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.5.1</version>
+				<configuration>
+					<source>1.8</source>
+					<target>1.8</target>
+				</configuration>
+			</plugin>
 			<plugin>
 				<groupId>org.eclipse.tycho</groupId>
 				<artifactId>tycho-surefire-plugin</artifactId>
 				<configuration>
-					<appArgLine>-pluginCustomization ${requirementsDirectory}/pluginCustomization.ini</appArgLine>
+					<appArgLine>-pluginCustomization
+						${requirementsDirectory}/pluginCustomization.ini</appArgLine>
 					<useUIThread>false</useUIThread>
-					<testSuite>org.jboss.tools.jmx.ui.bot.test</testSuite>
+					<testSuite>org.jboss.tools.as.jmx.ui.bot.itests</testSuite>
 					<testClass>${suiteClass}</testClass>
 					<skip>${skipITests}</skip>
 				</configuration>
