+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/projects/SimpleWebApp.war
