+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/resources/requirements.json
