+++ as/itests/org.jboss.tools.as.jmx.ui.bot.itests/src/org/jboss/tools/as/jmx/ui/bot/itests/server/connection/JMXServerConnectionAutoDeployDisableTest.java
@@ -8,7 +8,7 @@
  * Contributors: 
  * Red Hat, Inc. - initial API and implementation 
  ******************************************************************************/
-package org.jboss.tools.jmx.ui.bot.test.server.connection;
+package org.jboss.tools.as.jmx.ui.bot.itests.server.connection;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
