+++ as/itests/pom.xml
@@ -11,9 +11,10 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xs
 	<name>as.itests</name>
 	<packaging>pom</packaging>
 	<modules>
-		<module>org.jboss.tools.as.itests</module> 
+		<module>org.jboss.tools.as.itests</module>
 		<module>org.jboss.tools.as.management.itests</module>
 		<module>org.jboss.tools.as.ui.bot.itests</module>
+		<module>org.jboss.tools.as.jmx.ui.bot.itests</module>
 	</modules>
 
 	<properties>
