--- jmx/itests/pom.xml
@@ -1,18 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" 
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-		xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.jboss.tools</groupId>
-    <artifactId>jmx</artifactId>
-    <version>1.10.6-SNAPSHOT</version>
-  </parent>
-  <groupId>org.jboss.tools.jmx</groupId>
-  <artifactId>itests</artifactId>
-  <packaging>pom</packaging>
-  <name>jmx.itests</name>
-  	<modules>
-		<module>org.jboss.tools.jmx.ui.bot.test</module>
-	</modules>
-</project>
