+++ jmx/pom.xml
@@ -16,7 +16,6 @@
 		<module>plugins</module>
 		<module>tests</module>
 		<module>features</module>
-		<module>itests</module>
 		<module>test-framework</module>
 	</modules>
