+++ jmx/test-framework/org.jboss.tools.jmx.reddeer/pom.xml
@@ -9,6 +9,5 @@
 	</parent>
 	<groupId>org.jboss.tools.jmx.tests</groupId>
 	<artifactId>org.jboss.tools.jmx.reddeer</artifactId>
-	<version>1.10.100-SNAPSHOT</version>
 	<packaging>eclipse-plugin</packaging>
 </project>
